Fix permissions on scripts
The COPY statements creates files for the root user by default.

In case that the user running on the host does not have any group
or other permissons set (umask 077) this ends up in the script owned by
root but without any read or exec permission for the sdr user inside the docker container.

So explicitely set the file mode for script files.